requirements.txt
torch
numpy
pyfunctional
tqdm
jupyter
scikit-learn
torchvision
matplotlib
